ZOMG THREADING NIGHTMARE!
There is an issue when downloading multiple files at once (like a directory) where the listener backlog fills up and eventually rejects socket connections. If the processing is threaded then packets save out of order or cause file access errors as they try to save at the same time. Either thread the send or the receive...